Refactoring adding a project to phenotips out of the command that adds it to xBrowse
